corrected [compat]
For versions 0.x, the caret is interpreted differently: `"^0.1"` covers up to but not including `0.2`. Uncorrected, this forces use of a non-current version of `BFloats16.jl`  and causes other packages to be downgraded (including `CUDA.jl`).
I also bumped your patch version number. After merging this patch, you will need to tell Julia to register the update by adding this as a comment to the latest change  `@JuliaRegistrator register()`.  Clicking on the history message "Merge pull request .." opens the comments.